Rename and find npm path more elegantly
